Remove shared Radio
